move multiplexer to a separate module
