+++ tests/test_multiplexer.py
@@ -32,7 +32,8 @@ import pytest
 
 import aea
 from aea.configurations.base import PublicId
-from aea.mail.base import AEAConnectionError, Envelope, EnvelopeContext, Multiplexer
+from aea.mail.base import AEAConnectionError, Envelope, EnvelopeContext
+from aea.multiplexer import Multiplexer
 from aea.protocols.default.message import DefaultMessage
 from aea.protocols.default.serialization import DefaultSerializer
