+++ tests/test_packages/test_connections/test_local/test_search_services.py
@@ -22,15 +22,10 @@ import time
 
 import pytest
 
-from aea.helpers.search.models import (
-    Attribute,
-    Constraint,
-    ConstraintType,
-    DataModel,
-    Description,
-    Query,
-)
-from aea.mail.base import AEAConnectionError, Envelope, InBox, Multiplexer
+from aea.helpers.search.models import (Attribute, Constraint, ConstraintType,
+                                       DataModel, Description, Query)
+from aea.mail.base import AEAConnectionError, Envelope
+from aea.multiplexer import InBox, Multiplexer
 from aea.protocols.default.message import DefaultMessage
 from aea.protocols.default.serialization import DefaultSerializer
 
@@ -38,7 +33,8 @@ from packages.fetchai.connections.local.connection import LocalNode
 from packages.fetchai.protocols.fipa.message import FipaMessage
 from packages.fetchai.protocols.fipa.serialization import FipaSerializer
 from packages.fetchai.protocols.oef_search.message import OefSearchMessage
-from packages.fetchai.protocols.oef_search.serialization import OefSearchSerializer
+from packages.fetchai.protocols.oef_search.serialization import \
+    OefSearchSerializer
 
 from ....conftest import _make_local_connection
