+++ tests/test_packages/test_connections/test_p2p_libp2p/test_communication.py
@@ -28,7 +28,8 @@ from typing import Optional, Sequence
 import pytest
 
 from aea.crypto.fetchai import FetchAICrypto
-from aea.mail.base import Envelope, Multiplexer
+from aea.mail.base import Envelope
+from aea.multiplexer import Multiplexer
 from aea.protocols.default.message import DefaultMessage
 from aea.protocols.default.serialization import DefaultSerializer
