+++ tests/test_packages/test_connections/test_tcp/test_communication.py
@@ -25,7 +25,8 @@ import unittest.mock
 
 import pytest
 
-from aea.mail.base import Envelope, Multiplexer
+from aea.mail.base import Envelope
+from aea.multiplexer import Multiplexer
 from aea.protocols.default.message import DefaultMessage
 from aea.protocols.default.serialization import DefaultSerializer
